chore(DTFS2-6829): update coverEndDate in e2e test spec
