+++ e2e-tests/package.json
@@ -32,6 +32,7 @@
     "date-fns": "^2.30.0",
     "eslint": "^8.56.0",
     "eslint-plugin-cypress": "^2.15.1",
+    "lodash": "^4.17.21",
     "mongodb": "^6.3.0"
   },
   "engines": {
